searches for a value in an array of integers using the Linear search algorithm

+++ 0x1E-search_algorithms/0-linear.c
@@ -13,14 +13,14 @@
 
 int linear_search(int *array, size_t size, int value)
 {
-	size_t i;
+	int i;
 
 	if (array == NULL || size == 0)
 		return (-1);
 
-	for (i = 0; i < size; i++)
+	for (i = 0; i < (int)size; i++)
 	{
-		printf("value checked array[%ld] = [%d]\n", i, array[i]);
+		printf("value checked array[%d] = [%d]\n", i, array[i]);
 		if (array[i] == value)
 		{
 			return (i);
